Backend/requirements.txt
fastapi
uvicorn
openai
redis
python-dotenv
httpx
pydantic
langdetect
pymongo
requests
google-generativeai
